Third Commit
Changed it up to be much more crisp.  I hated the floaty movement and
didn't think of a way to fix it, but here we are, I did it!  Still more
I want to do, such as fixing the variables up some, and improving upon
the wall jump a lot more.